soft/requirements.txt
Pillow
fonttools
pyyaml
pyserial
